Added server logic
Added server logic directory with script.js.
This is extended version which also gives a wonderful opportunity to
work with tags. Works really well.
Tags also include colors that is why it will be looking great with
colored dots on the signs. Looking forward to including this code to
server logic.
Nevertheless html part should be rebuilt fully :(